backend/requirements.txt
fastapi
uvicorn
python-multipart
pymupdf
moviepy<2.0.0
numpy
pillow
websockets
